Refactoring in performance test (may be build, but doesn't work)
