+++ dbms/programs/performance-test/CMakeLists.txt
@@ -3,7 +3,13 @@ add_library (clickhouse-performance-test-lib ${LINK_MODE}
   StopConditionsSet.cpp
   TestStopConditions.cpp
   TestStats.cpp
+  ConfigPreprocessor.cpp
   PerformanceTest.cpp
+  PerformanceTestInfo.cpp
+  executeQuery.cpp
+  applySubstitutions.cpp
+  ReportBuilder.cpp
+  PerformanceTestSuite.cpp
 )
 target_link_libraries (clickhouse-performance-test-lib PRIVATE dbms clickhouse_common_io clickhouse_common_config ${Boost_PROGRAM_OPTIONS_LIBRARY})
 target_include_directories (clickhouse-performance-test-lib SYSTEM PRIVATE ${PCG_RANDOM_INCLUDE_DIR})
